# maefa-store/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: API Maefa Store
  version: 1.0.0
  description: |
    API REST du serveur Maefa Store (Node.js + Express). Toutes les réponses sont en JSON.
    Ouvrir ce fichier dans https://editor.swagger.io pour une page interactive.

    Accès :
    - **gérante** : en-tête `x-admin-pin` (jeton signé obtenu par `POST /api/admin/login`) ;
    - **cliente connectée** : en-tête `Authorization: Bearer <jeton>` (obtenu par `/api/auth/verify`) ;
    - **livreur** : jeton secret dans l'adresse (`/api/driver/{token}`, `/api/tour/{token}`).
servers:
  - url: https://maefa-store.onrender.com
  - url: http://localhost:8787
components:
  securitySchemes:
    admin: { type: apiKey, in: header, name: x-admin-pin }
    session: { type: http, scheme: bearer }
...
paths:
...
  /api/auth/logout:
    post:
      tags: [Comptes clientes]
      summary: 'Se déconnecter'
      security: [{ session: [] }]
      responses:
        '200': { description: OK }
  /api/me:
    get:
      tags: [Comptes clientes]
      summary: 'Profil de la cliente'
      security: [{ session: [] }]
      responses:
        '200': { description: OK }
    patch:
      tags: [Comptes clientes]
      summary: 'Modifier le profil et les adresses enregistrées'
      security: [{ session: [] }]
      responses:
        '200': { description: OK }
    delete:
      tags: [Comptes clientes]
      summary: "Droit à l'effacement : supprimer le compte, ses sessions et son historique"
      security: [{ session: [] }]
      responses:
        '200': { description: Compte supprimé }
        '401': { description: Non connectée }
        '409': { description: Une commande est en cours }
  /api/me/export:
    get:
      tags: [Comptes clientes]
      summary: "Droit d'accès (loi 2008-12) : toutes les données gardées sur la cliente"
      security: [{ session: [] }]
      responses:
        '200': { description: Fichier JSON (profil, adresses, commandes), sans le code secret }
        '401': { description: Non connectée }
  /api/me/pin:
    post:
      tags: [Comptes clientes]
      summary: 'Choisir un code personnel'
      security: [{ session: [] }]
      responses:
        '200': { description: OK }
  /api/me/orders:
    post:
      tags: [Comptes clientes]
      summary: 'Rattacher des commandes au compte'
      security: [{ session: [] }]
      responses:
        '200': { description: OK }
  /api/admin/customers:
    get:
      tags: [Comptes clientes]
      summary: 'Liste des clientes'
      security: [{ admin: [] }]
      responses:
        '200': { description: OK }
        '403': { description: Accès refusé }
  /api/admin/customers/{phone}/reset-pin:
    post:
      tags: [Comptes clientes]
      summary: "Réinitialiser le code d'une cliente"
      parameters:
        - { name: phone, in: path, required: true, schema: { type: string } }
      security: [{ admin: [] }]
      responses:
        '200': { description: OK }
        '403': { description: Accès refusé }
...